yume_lab/requirements-test.txt
# Testing dependencies for Yume dream system

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0

# Async testing
asyncio>=3.4.3

# For Redis testing
redis>=4.6.0
fakeredis>=2.19.0

# Core dependencies (should match main project)
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.8.0

# Transformers for CLIP
transformers>=4.30.0

openai
# OpenAI CLIP
git+https://github.com/openai/CLIP.git

# Coverage visualization
coverage[toml]>=7.2.0

# Watch mode
ptw
